make spline work better at edges
funny story: it wasn't just premature optimization!  at least, not entirely.  it turns out it's necessary to pad the edges of the array to make this convolution-spline be well-behaved at the edges, tho the way I did it in the last commit wasn't quite right.  doing linear extrapolation at the edges ensures that it actually remains fairly linear thruout the defined part of the spline.