vectorize spline inversion
I changed the spline inversion to be pure numpy because I need to vectorize it and I can't find a way to do vectorized gradient calculations in pytorch (vmap almost does what I want but doesn't work for indexing operations which I need for my, uh, prematurely optimized spline calculation).  so instead I just manually calculated the gradients (it wasn't very complicated) and manually implemented Newton–Raphson iteration (again, pretty simple to do, albeit not as likely to converge as Levenberg–Marquardt).  I moved the spline node padding to the spline constructor and moved the digitization stuff to its own little function to help reduce duplicate code, and whamo!  it's fully vectorized!  and it's still way too slow.  I'm actually not sure if it's faster than the previus verison or not.  I have to beleve it is, since it at least no longer has that terrible for-loop, but either way I need to do some kind of interpolation to make this finish in a timely manner.