add regularization term
I added a term to tell it to find smoother things.  I'm not sure if it's working.  it seems to be okay with some forms of skew that I would not consider to be zero-curvature (specificly d²z/dxdy).  but more importantly the results don't make any sense in a way that makes me think my cost function is wrong.