reduce redundant calculations in bicubic function
rather than calculating every piece of the piecewise function for every node in the spline, I realized that the way I've set it up makes it trivial to only calculate the one relevant section of the spline.  and the nice thing about this is that it actually makes extrapolation just work.  so now the edges of my inverted images look just a little bit better.