calculate residuals in warpd, not flat, space
I added some cmoplexty to the way residuals are calculated so that it uses the distance between the points and the line in warped space, not flattened space.  approximately, anyway; it takes the residual as calculated previusly and scales it based on the inverse jacobian at that point.  this should remove the incentive for it to simply shrink the image to get a better fit.

the bad news is that it doesn't work at all.  it still shrinks the image where it can and I don't understand why.  also, it's extremely slow now.  I think there's hope, tho.  first I need to debug this because I really don't se why it would be shrinking them now unless I've done something rong.  also, I think I can speed it up by using linalg.lstsq instead of linalg.inv.  also, it looks like I can probably remove one of the calls to sqrt in fit_arc.  so stand by for better news.